requirements-integration.txt
ufal.udpipe
fasttext
classla
stanza
